Track the working directory in the bootstrap command.
