+++ agent/bootstrap_shell_pty_windows.go
@@ -1,0 +1,5 @@
+package agent
+
+func (b Bootstrap) shellPTYStart(c *os.Cmd) {
+	// no-op
+}
